requirements.txt
py-tgcalls[pyrogram]==2.2.11
pyrogram==2.0.106
TgCrypto
yt-dlp
python-dotenv
Pillow
aiohttp
